+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+NOT RELEASED YET
+=================
+- Add create connection latencies metrics [#54](https://github.com/pfreixes/emcache/pull/54)
+
 0.4.0
 ==========
 - Stable version